Find available indicator code names in the Wittgenstein Centre Human Capital Data Explorer

Find available indicator code names in the Wittgenstein Centre Human Capital Data Explorer

Usage

find_indicator(x)

Arguments

x

Character string on key word or name related to indicator of potential interest.

Value

A subset of the wic_indicators data frame with one or more of the indicator, description or definition columns matching the keyword given to x. Use the result in the indicator column to input to the get_wcde function for downloading data.

Examples

find_indicator("education")
#> # A tibble: 11 × 6
#>    indicator description         `wcde-v3` `wcde-v2` `wcde-v1` definition_latest
#>    <chr>     <chr>               <chr>     <chr>     <chr>     <chr>            
#>  1 bpop      Population Size by… past-ava… past-ava… past-ava… De facto populat…
#>  2 bprop     Educational Attain… past-ava… past-ava… past-ava… De facto populat…
#>  3 easfr     Age-Specific Ferti… projecti… projecti… NA        Number of births…
#>  4 eassr     Age-Specific Survi… projecti… projecti… projecti… Survivorship of …
#>  5 emacb     Mean Age at Childb… projecti… NA        NA        The mean age of …
#>  6 epop      Population Size by… past-ava… past-ava… past-ava… De facto populat…
#>  7 etfr      Total Fertility Ra… projecti… projecti… projecti… The average numb…
#>  8 ggapedu15 Gender Gap in Educ… past-ava… past-ava… past-ava… The difference b…
#>  9 ggapedu25 Gender Gap in Educ… past-ava… past-ava… past-ava… The difference b…
#> 10 netedu    Net Migration Flow… projecti… NA        NA        The number of im…
#> 11 prop      Educational Attain… past-ava… past-ava… past-ava… De facto populat…
find_indicator("migr")
#> # A tibble: 4 × 6
#>   indicator description          `wcde-v3` `wcde-v2` `wcde-v1` definition_latest
#>   <chr>     <chr>                <chr>     <chr>     <chr>     <chr>            
#> 1 emi       Emigration Flows     projecti… NA        NA        The number of em…
#> 2 imm       Immigration Flows    projecti… NA        NA        The number of im…
#> 3 net       Net Migration Flows  past-ava… past-ava… past-ava… The number of im…
#> 4 netedu    Net Migration Flows… projecti… NA        NA        The number of im…
find_indicator("fert")
#> # A tibble: 4 × 6
#>   indicator description          `wcde-v3` `wcde-v2` `wcde-v1` definition_latest
#>   <chr>     <chr>                <chr>     <chr>     <chr>     <chr>            
#> 1 asfr      Age-Specific Fertil… past-ava… past-ava… past-ava… Number of births…
#> 2 easfr     Age-Specific Fertil… projecti… projecti… NA        Number of births…
#> 3 etfr      Total Fertility Rat… projecti… projecti… projecti… The average numb…
#> 4 tfr       Total Fertility Rate past-ava… past-ava… past-ava… The average numb…