Indicators used in the Wittgenstein Data Explorer

A data set containing the indicator codes, names and further details used in the Wittgenstein Data Explorer

wic_indicators

Format

A data frame with 31 rows and 8 variables, including:

indicator

Short name of indicator to be used in the indicator argument of the get_wcde() function

description

Brief description of indicator

age

Availability of indicator by five-year age groups

bage

Availability of indicator by broad age groups

sage

Availability of indicator with a new born age group

sex

Availability of indicator by sex

edu

Availability of indicator by education

period

Indicator is a period (flow)

past

Availability of past data for indicator

definition

Full definition for indicator

Source

http://dataexplorer.wittgensteincentre.org/wcde-v2/