Download data from the Wittgenstein Centre Human Capital Data Explorer Data Explorer

Downloads data from the Wittgenstein Centre Human Capital Data Explorer. Requires a working internet connection.

get_wcde(
  indicator = "pop",
  scenario = 2,
  country_code = NULL,
  country_name = NULL,
  include_scenario_names = FALSE
)

Arguments

indicator
One character string based on the indicator column in the wic_indicators data frame, representing the variable to be downloaded.

scenario
Vector of length one or more with numbers corresponding the scenarios. See details for more information. Defaults to 2 for the SSP2 Medium scenario.

country_code
Vector of length one or more of country numeric codes based on ISO 3 digit numeric values.

country_name
Vector of length one or more of country names. The corresponding country code will be guessed using the countrycodes package.

include_scenario_names
Logical vector of length one to indicate if to include additional columns for scenario names and short names. FALSE by default.

Value

A tibble with the data selected.

Details

If not country_name or country_code is provided data for all countries and regions are downloaded. A full list of available countries and regions can be found in the wic_locations data frame.

indicator must be set to a value in the first column in the table below of available demographic indicators:

indicator	Indicator Description

pop	Population Size (000's)

bpop	Population Size by Broad Age (000's)

epop	Population Size by Education (000's)

prop	Educational Attainment Distribution

bprop	Educational Attainment Distribution by Broad Age

growth	Average Annual Growth Rate

nirate	Average Annual Rate of Natural Increase

sexratio	Sex Ratio

mage	Population Median Age

tdr	Total Dependency Ratio

ydr	Youth Dependency Ratio

odr	Old-age Dependency Ratio

ryl15	Age When Remaining Life Expectancy is Below 15 years

pryl15	Proportion of Population with a Remaining Life Expectancy below 15 Years

mys	Mean Years of Schooling by Age

bmys	Mean Years of Schooling by Broad Age

ggapmys15	Gender Gap in Mean Years Schooling (15+)

ggapmys25	Gender Gap in Mean Years Schooling (25+)

ggapedu15	Gender Gap in Educational Attainment (15+)

ggapedu25	Gender Gap in Educational Attainment (25+)

tfr	Total Fertility Rate

etfr	Total Fertility Rate by Education

asfr	Age-Specific Fertility Rate

easfr	Age-Specific Fertility Rate by Education

cbr	Crude Birth Rate

macb	Mean Age at Childbearing

e0	Life Expectancy at Birth

cdr	Crude Death Rate

assr	Age-Specific Survival Ratio

eassr	Age-Specific Survival Ratio by Education

net	Net Migration

See wic_indicators data frame for more details.

scenario must be set to one or values in the first column table below of the available future scenarios:

scenario	description

1	Rapid Development (SSP1)

2	Medium (SSP2)

3	Stalled Development (SSP3)

21	Medium - Zero Migration (SSP2 - ZM)

22	Medium - Double Migration (SSP2 - DM)

See wic_scenarios data frame for more details.

Examples

# \donttest{
# SSP2 tfr for Austria and Bulgaria
get_wcde(indicator = "tfr", country_code = c(40, 100))

#> # A tibble: 60 x 5
#>    scenario name     country_code period      tfr
#>       <dbl> <chr>           <dbl> <chr>     <dbl>
#>  1        2 Austria            40 1950-1955  2.1 
#>  2        2 Bulgaria          100 1950-1955  2.53
#>  3        2 Austria            40 1955-1960  2.57
#>  4        2 Bulgaria          100 1955-1960  2.3 
#>  5        2 Austria            40 1960-1965  2.78
#>  6        2 Bulgaria          100 1960-1965  2.22
#>  7        2 Austria            40 1965-1970  2.57
#>  8        2 Bulgaria          100 1965-1970  2.13
#>  9        2 Austria            40 1970-1975  2.04
#> 10        2 Bulgaria          100 1970-1975  2.16
#> # ... with 50 more rows

# SSP1 and SSP2 life expectancy for Vietnam and United Kingdom (guessing the country codes)
get_wcde(scenario = c(1, 2), indicator = "e0", country_name = c("Vietnam", "UK"))

#> # A tibble: 240 x 6
#>    scenario name                               country_code sex    period     e0
#>       <dbl> <chr>                                     <dbl> <chr>  <chr>   <dbl>
#>  1        1 Viet Nam                                    704 Male   1950-1~  50.8
#>  2        1 United Kingdom of Great Britain a~          826 Male   1950-1~  66.8
#>  3        1 Viet Nam                                    704 Female 1950-1~  56.5
#>  4        1 United Kingdom of Great Britain a~          826 Female 1950-1~  71.9
#>  5        1 Viet Nam                                    704 Male   1955-1~  54  
#>  6        1 United Kingdom of Great Britain a~          826 Male   1955-1~  67.7
#>  7        1 Viet Nam                                    704 Female 1955-1~  60.7
#>  8        1 United Kingdom of Great Britain a~          826 Female 1955-1~  73.4
#>  9        1 Viet Nam                                    704 Male   1960-1~  56.8
#> 10        1 United Kingdom of Great Britain a~          826 Male   1960-1~  68  
#> # ... with 230 more rows

# SSP1 and SSP3 population by education for all countries
get_wcde(scenario = c(1, 3), indicator = "tfr")

#> # A tibble: 13,740 x 5
#>    scenario name                     country_code period      tfr
#>       <dbl> <chr>                           <dbl> <chr>     <dbl>
#>  1        1 Bulgaria                          100 1950-1955  2.53
#>  2        1 Myanmar                           104 1950-1955  6   
#>  3        1 Burundi                           108 1950-1955  6.8 
#>  4        1 Belarus                           112 1950-1955  2.61
#>  5        1 Cambodia                          116 1950-1955  6.95
#>  6        1 Algeria                            12 1950-1955  7.28
#>  7        1 Cameroon                          120 1950-1955  5.49
#>  8        1 Canada                            124 1950-1955  3.64
#>  9        1 Cape Verde                        132 1950-1955  6.57
#> 10        1 Central African Republic          140 1950-1955  5.52
#> # ... with 13,730 more rows

# }